wifi: ath10k: Delay the unmapping of the buffer
We are seeing a corner case where host receiving copy completion for WMI
command though copy engine is processing it. Once host receives the
copy completion, host is unmapping corresponding memory which results
in SMMU fault. To avoid such conditions though host receives copy
completion, as a work around we are adding a delay to unmap the memory
for WMI end point.

Tested-on: WCN3990 hw1.0 SNOC WLAN.HL.2.0-01387-QCAHLSWMTPLZ-1

Tested-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Youghandhar Chintala <quic_youghand@quicinc.com>
Patchwork-Id: 12944385
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>